+++ src/lib/pass-event.js
@@ -1,0 +1,9 @@
+/**
+ * @param {EventEmitter} sourceEmitter
+ * @param {EventEmitter} destEmitter
+ * @param {String} sourceEvent
+ * @param {String} destEvent
+ */
+export default function passEvent(sourceEmitter, destEmitter, sourceEvent, destEvent) {
+	sourceEmitter.on(sourceEvent, (...args) => destEmitter.emit(destEvent, ...args))
+}